requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
cupy-cuda12x>=12.0.0  # For CUDA 12.x, adjust based on your CUDA version
# Alternative for CUDA 11.x: cupy-cuda11x>=11.0.0
# For CPU-only: comment out cupy line above

